INDY-2068: Review fixes
